requirementso.txt
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
autograd==1.3
autoscraper==1.1.12
backcall==0.2.0
bkcharts==0.2
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.12.5
colorama==0.4.4
comtypes==1.1.9
confuse==1.4.0
converter==1.0.0
cssselect==1.1.0
cycler==0.10.0
cytoolz==0.11.0
decorator==4.4.2
Delorean==1.0.0
Deprecated==1.2.12
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
future==0.18.2
gast==0.2.2
grpcio==1.32.0
gunicorn==20.1.0
HeapDict==1.0.1
hiredis==2.0.0
htmlmin==0.1.12
httplib2==0.18.1
humanize==3.1.0
ImageHash==4.2.0
imbalanced-learn==0.7.0
inflection==0.5.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
json5==0.9.5
lesscpy==0.13.0
lightgbm==3.2.1
locket==0.2.1
mccabe==0.6.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
numpy==1.19.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pafy==0.5.5
parso==0.8.1
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
phik==0.11.2
pickle-mixin==1.0.2
pickleshare==0.7.5
pkginfo==1.7.0
ply==3.11
proglog==0.1.9
prompt-toolkit==3.0.14
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycrypto==2.6.1
pycurl==7.43.0.6
pydub==0.24.1
Pygments==2.7.4
pynndescent==0.5.2
pyreadline==2.1
pytest==6.2.3
python-dateutil==2.8.1
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.2
QtPy==1.9.0
redis==3.5.3
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.2
scipy==1.6.3
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
sklego==0.0
statsmodels==0.12.0
tabulate==0.8.9
tangled-up-in-unicode==0.0.7
terminado==0.9.4
terminaltables==3.1.0
Theano==1.0.4
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
umap-learn==0.5.1
uritemplate==3.0.1
visions==0.6.0
wcwidth==0.2.5
webencodings==0.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xgboost==1.3.1
xlwt==1.3.0
zict==2.0.0
zope.event==4.5.0